+++ app/components/scoreboard_summary_component/scoreboard_summary_component.html.erb
@@ -33,7 +33,7 @@
         </div>
       </div>
       <div class="col-12 col-md-6">
-        <%= render PodiumComponent.new(podium: podium, user: user, classes: 'bg-white rounded p-4') %>
+        <%= render Scoreboards::PodiumComponent.new(podium: podium, user: user, classes: 'bg-white rounded p-4') %>
       </div>
     </div>
     <%= render TitledSectionComponent.new(id: 'scoreboard-recent-activity', classes: 'mt-4') do |section| %>
@@ -71,7 +71,7 @@
         <div class="col-lg card-deck actions">
           <div class="card">
             <div class="card-body">
-                <%= render PodiumComponent.new(podium: podium) %>
+                <%= render Scoreboards::PodiumComponent.new(podium: podium) %>
             </div>
           </div>
         </div>
