+++ app/components/scoreboards/podium_component/podium_component.html.erb
