+++ app/components/scoreboards/podium_component/podium_component.scss
@@ -1,4 +1,4 @@
-.podium-component {
+.scoreboards-podium-component {
   .bar {
     width: 95%;
     margin-top: 10px;
