+++ spec/components/scoreboards/podium_component_spec.rb
@@ -2,12 +2,11 @@
 
 require 'rails_helper'
 
-RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
+RSpec.describe Scoreboards::PodiumComponent, :include_url_helpers, type: :component do
   let(:scoreboard) { create :scoreboard }
   let(:school) { create :school, scoreboard: scoreboard }
   let(:podium) { Podium.create(school: school, scoreboard: school.scoreboard) }
-  let(:all_params) { { podium: podium, classes: 'my-class', id: 'my-id' } }
-  let(:params) { all_params }
+  let(:params) { { podium: podium, classes: 'my-class', id: 'my-id' } }
 
   # Avoids problem with showing national placing. National Scoreboard only runs from
   # 1st Sept to 31st Jul.
@@ -22,7 +21,7 @@ RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
   end
 
   let(:html) do
-    render_inline(PodiumComponent.new(**params))
+    render_inline(described_class.new(**params))
   end
 
   shared_examples 'a podium including school' do
@@ -69,22 +68,22 @@ RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
   context 'with all params' do
     let(:school) { create :school, :with_points, score_points: 50, scoreboard: scoreboard }
 
-    it { expect(html).to have_selector('div.podium-component') }
+    it { expect(html).to have_selector('div.scoreboards-podium-component') }
 
     it 'adds specified classes' do
-      expect(html).to have_css('div.podium-component.my-class')
+      expect(html).to have_css('div.scoreboards-podium-component.my-class')
     end
 
     it 'adds specified id' do
-      expect(html).to have_css('div.podium-component#my-id')
+      expect(html).to have_css('div.scoreboards-podium-component#my-id')
     end
   end
 
   context 'when there is another school on the podium' do
-    let!(:other_school) { create :school, :with_points, score_points: 50, scoreboard: scoreboard }
-
     context 'when school is in first place' do
       let(:school) { create :school, :with_points, score_points: 60, scoreboard: scoreboard }
+
+      before { create :school, :with_points, score_points: 50, scoreboard: scoreboard }
 
       it { expect(html).to have_content("You are in 1st place on the #{scoreboard.name} scoreboard") }
       it { expect(html).to have_content('and 1st place nationally') }
@@ -96,6 +95,8 @@ RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
     context 'when in second place' do
       let(:school) { create :school, :with_points, score_points: 30, scoreboard: scoreboard }
 
+      before { create :school, :with_points, score_points: 50, scoreboard: scoreboard }
+
       it { expect(html).to have_content("You are in 2nd place on the #{scoreboard.name} scoreboard") }
       it { expect(html).to have_content('and 2nd place nationally') }
 
@@ -104,9 +105,9 @@ RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
     end
 
     context 'when in second place nationally' do
-      let!(:other_school) { create :school, :with_points, score_points: 50, scoreboard: create(:scoreboard) }
-
       let(:school) { create :school, :with_points, score_points: 30, scoreboard: scoreboard }
+
+      before { create :school, :with_points, score_points: 50, scoreboard: create(:scoreboard) }
 
       it { expect(html).to have_content("You are in 1st place on the #{scoreboard.name} scoreboard") }
       it { expect(html).to have_content('and 2nd place nationally') }
@@ -117,6 +118,8 @@ RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
 
     context "when school doesn't have any points" do
       let(:school) { create :school, scoreboard: scoreboard }
+
+      before { create :school, :with_points, score_points: 50, scoreboard: scoreboard }
 
       it_behaves_like 'a podium including school'
       it_behaves_like 'a podium with no points message'
@@ -132,7 +135,7 @@ RSpec.describe PodiumComponent, type: :component, include_url_helpers: true do
   end
 
   context 'with no podium' do
-    let(:params) { all_params.except(:podium) }
+    let(:params) { { classes: 'my-class', id: 'my-id' } }
 
     it "doesn't render" do
       expect(html.to_s).to be_blank
